requirements.txt
gradio
torch
torchvision
numpy
pillow
matplotlib
captum
opencv-python
